Make rendering more powerful by including HTML context and arbitrary key
This allows folks to get more powerful rendering out of their templates.
It adds:

If you set ``:key: foo`` on the directive,
you will have a top-level ``key`` variable.
This can be used to render a subset of the data in the template.

If you set `:include_context:` on the directive,
you template will also have access to the builder's ``html_context`` while rendering.
This will be in the global template context, not inside ``data``.